Update the dewclaw-deps package in-place in /tmp
I don't see a reason for keeping multiple old version around there. /tmp
is on a tmpfs, but still it seems unneccessary to risk filling it up if
using the dewclaw --reload flag (which avoids reboot / reset of /tmp).

While at it, use ${package_name} to keep a bit more DRY.